Updated game field representations, added .gitignore
